hw2 commit which create logger with colors
