+++ config/default.json
@@ -1,0 +1,4 @@
+{
+  "COLORS_ENABLED": "0",
+  "LOG_LEVEL": "warn"
+}
